# src/job_pricing/docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Dynamic Job Pricing Engine API
  version: 1.0.0
  description: |
    Production-ready API for AI-powered compensation intelligence platform.

    ## Features
    - Job pricing with confidence scoring
    - Mercer job library integration
    - SSG skills extraction
    - Market data aggregation
    - Historical trends analysis

    ## Authentication
    All endpoints require JWT Bearer token authentication except /auth endpoints.

    ## Rate Limiting
    - Free Tier: 100 requests/hour
    - Standard Tier: 1000 requests/hour
    - Enterprise Tier: Unlimited

  contact:
    name: API Support
    email: api-support@company.com
  license:
    name: Proprietary
servers:
  - url: https://api.jobpricing.company.com/v1
    description: Production server
  - url: https://staging-api.jobpricing.company.com/v1
    description: Staging server
  - url: http://localhost:8000/v1
    description: Development server
...
security:
  - bearerAuth: []
paths:
...
  /job-pricing/{request_id}:
    get:
      tags:
        - Job Pricing
      summary: Get pricing result
      description: Retrieve pricing result for a specific request
      parameters:
        - name: request_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Pricing result retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PricingResult'
        '404':
          description: Request not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /mercer/jobs/{job_code}:
    get:
      tags:
        - Mercer Integration
      summary: Get Mercer job details
      description: Retrieve detailed information for a specific Mercer job
      parameters:
        - name: job_code
          in: path
          required: true
          schema:
            type: string
          example: HRM.04.005.M50
      responses:
        '200':
          description: Job details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MercerJobDetail'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PricingResult:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            request_id:
              type: string
              format: uuid
            job_summary:
              type: object
              properties:
                job_title:
                  type: string
                location:
                  type: string
                internal_grade:
                  type: string
                employment_type:
                  type: string
            mercer_mapping:
              type: object
              properties:
                job_code:
                  type: string
                job_title:
                  type: string
                position_class:
                  type: integer
                confidence:
                  type: number
                  format: float
            salary_recommendation:
              type: object
              properties:
                currency:
                  type: string
                  example: SGD
                period:
                  type: string
                  example: annual
                recommended_range:
                  type: object
                  properties:
                    min:
                      type: number
                      format: float
                      example: 72000
                    max:
                      type: number
                      format: float
                      example: 95000
                    target:
                      type: number
                      format: float
                      example: 82000
                full_distribution:
                  type: object
                  properties:
                    p10:
                      type: number
                      format: float
                    p25:
                      type: number
                      format: float
                    p50:
                      type: number
                      format: float
                    p75:
                      type: number
                      format: float
                    p90:
                      type: number
                      format: float
            confidence_metrics:
              type: object
              properties:
                overall_score:
                  type: number
                  format: float
                  minimum: 0
                  maximum: 100
                confidence_level:
                  type: string
                  enum: [High, Medium, Low]
                recommendation:
                  type: string
            data_sources:
              type: object
              additionalProperties:
                type: object
                properties:
                  weight_applied:
                    type: number
                    format: float
                  p50:
                    type: number
                    format: float
                  sample_size:
                    type: integer
                  data_date:
                    type: string
                    format: date
            explanation:
              type: object
              properties:
                summary:
                  type: string
                key_factors:
                  type: array
                  items:
                    type: string
                considerations:
                  type: array
                  items:
                    type: string
            alternative_scenarios:
              type: array
              items:
                type: object
                properties:
                  scenario:
                    type: string
                  range:
                    type: object
                    properties:
                      min:
                        type: number
                      max:
                        type: number
                  use_case:
                    type: string
            created_at:
              type: string
              format: date-time
...
    MercerJobDetail:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            job_code:
              type: string
            job_title:
              type: string
            family:
              type: string
            subfamily:
              type: string
            career_level:
              type: string
            position_class:
              type: integer
            job_description:
              type: string
            typical_titles:
              type: array
              items:
                type: string
            ipe_factors:
              type: object
              properties:
                impact:
                  type: object
                  properties:
                    min:
                      type: integer
                    max:
                      type: integer
                communication:
                  type: object
                  properties:
                    min:
                      type: integer
                    max:
                      type: integer
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: INVALID_INPUT
            message:
              type: string
              example: Validation error
            details:
              type: object
        metadata:
          type: object
          properties:
            timestamp:
              type: string
              format: date-time
            request_id:
              type: string
              format: uuid